requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
mysql-connector-python==8.0.33
protobuf==3.20.3
pytz==2023.3
sqlparse==0.4.4
